refactored popup.js
