--- content/ui/home.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <div id="browsingTime">
-        <h1>Ping kids mode!</h1>
-    </div>
-    <div id="browsingTimeOver" style="display: none;">
-        <h1>Your browsing time is over. Go back to study from your books.</h1>
-    </div>
-</body>
-<script src = "../content_scripts/popup.js" type="module"></script>
-</html>
